refactor: split components from ProjectEditForm
- CaptionsField
- CaptionField
