+++ src/renderer/components/ProjectEditForm.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Debug from 'debug';
-import { debounce } from 'throttle-debounce';
 
 import { connect } from 'react-redux';
+import { debounce } from 'throttle-debounce';
 
 import { updateProject } from '../actions/manager';
 
 import Player from './Player';
+import CaptionsField from './ProjectEditForm/CaptionsField';
 
 const debug = Debug('fabnavi:jsx:ProjectEditForm');
 
@@ -307,128 +308,6 @@ class ProjectEditForm extends React.Component {
     }
 }
 
-const CaptionsField = ({ figures, handleCaptionsChange, onAddCaptionButtonClick }) => {
-    return (
-        <div className='field_captions'>
-            <style jsx>{`
-                .field_captions {
-                    margin-left: 30px;
-                    overflow-y: scroll;
-                    height: 718px;
-                }
-                .edit {
-                    background-color: #C4C4C4;
-                    color: black;
-                    font-size: 20px;
-                    margin-top: 0;
-                    margin-bottom: 10px;
-                }
-                .field_caption_wrapper {
-                    margin-bottom: 20px;
-                }
-                .field_caption {
-                    margin-left: 100px;
-                }
-                .field_caption span {
-                    display: inline-block;
-                    width: 70px;
-                    margin: 0 20px;
-                }
-                .field_caption span:nth-child(4) {
-                    margin: 0;
-                }
-                .field_captions_index {
-                    margin-bottom: -24px;
-                }
-                ul {
-                    margin: 0;
-                    padding: 0;
-                }
-                .addCaptionButton {
-                    margin-left: 100px;
-                }
-            `}</style>
-            <p className="edit">
-                Captions
-            </p>
-            <div className='field_caption'>
-                <span style={{ marginLeft: '0' }}>start(sec)</span>
-                <span>end(sec)</span>
-                <span style={{ width: '200px' }}>text</span>
-                <span>destroy?</span>
-            </div>
-            {
-                figures.map((figure, figureIndex) => {
-                    return (
-                        <div className='field_caption_wrapper' key={`figure_${figureIndex}_captions`}>
-                            <div className='field_captions_index'>Image: {`${figureIndex + 1}`}</div>
-                            <ul>
-                                {
-                                    figure.captions.map((caption, index) => {
-                                        return (
-                                            <CaptionField
-                                                caption={caption}
-                                                index={index}
-                                                figureIndex={figureIndex}
-                                                handleCaptionsChange={handleCaptionsChange}
-                                                key={`caption_0_${index}`}
-                                            />
-                                        )
-                                    })
-                                }
-                            </ul>
-                            <button className="addCaptionButton" onClick={onAddCaptionButtonClick} data-index={figureIndex}>add caption</button>
-                        </div>
-                    )
-                })
-            }
-        </div>
-    )
-}
-
-CaptionsField.propTypes = {
-    figures: PropTypes.array,
-    handleCaptionsChange: PropTypes.func,
-    onAddCaptionButtonClick: PropTypes.func
-}
-
-const CaptionField = ({ caption, index, figureIndex, handleCaptionsChange }) => {
-    return (
-        <li className='field_caption' onChange={handleCaptionsChange} data-figure-index={figureIndex} data-index={index}>
-            <style jsx>{`
-                .field_caption {
-                    margin-left: 100px;
-                }
-                .field_caption span {
-                    display: inline-block;
-                    width: 70px;
-                    margin: 0 20px;
-                }
-                .field_caption input {
-                    margin: 0 18px;
-                }
-                .field_caption input[type="number"] {
-                    width: 70px;
-                }
-                .field_caption input[type="text"] {
-                    width: 200px;
-                }
-            `}</style>
-            <input name='id' data-index={index} type="hidden" defaultValue={caption.id || null}/>
-            <input name='start_sec' min="0" style={{ marginLeft: '0' }} data-index={index} type="number" defaultValue={caption.start_sec}/>
-            <input name='end_sec' min="0" data-index={index} type="number" defaultValue={caption.end_sec}/>
-            <input name='text' data-index={index} type="text" defaultValue={caption.text}/>
-            <input name='_destroy' data-index={index} type="checkbox" defaultChecked={false}/>
-        </li>
-    )
-};
-
-CaptionField.propTypes = {
-    caption: PropTypes.object,
-    index: PropTypes.number,
-    figureIndex: PropTypes.number,
-    handleCaptionsChange: PropTypes.func
-}
 
 ProjectEditForm.propTypes = {
     project: PropTypes.object,
